refactor: remove io mod. and move dot to topology
