+++ cspell.json
@@ -109,6 +109,7 @@
         "adrs",
         "arange",
         "asdict",
+        "asdot",
         "cano",
         "celltoolbar",
         "codacy",
